+++ mininterface/_lib/future_compatibility.py
@@ -1,6 +1,9 @@
 
-from typing import Literal
+from typing import Annotated, Literal
 
 
 def literal(c):
     return Literal[*c]
+
+def spread_annotated(obj, annotations):
+    return Annotated[obj, *annotations]