+++ mininterface/_lib/start.py
@@ -1,23 +1,23 @@
 # Starting and maintaining a program, using mininterface, in the system.
 import sys
-from collections import defaultdict
-from dataclasses import is_dataclass
+# from collections import defaultdict
+# from dataclasses import is_dataclass
 from pathlib import Path
 from subprocess import run
 from typing import Optional, Type
-from warnings import warn
+# from warnings import warn
 
 from .._mininterface import Mininterface
-from ..cli import Command, SubcommandPlaceholder
-from ..exceptions import DependencyRequired
+# from ..cli import Command, SubcommandPlaceholder
+# from ..exceptions import DependencyRequired
 from ..interfaces import get_interface
-from ..tag import Tag
-from .form_dict import DataClass, EnvClass, TagDict, dataclass_to_tagdict
+# from ..tag import Tag
+# from .form_dict import DataClass, EnvClass, TagDict, dataclass_to_tagdict
 
-try:
-    from .cli_parser import parse_cli
-except DependencyRequired as e:
-    parse_cli = e
+# try:
+#     from .cli_parser import parse_cli
+# except DependencyRequired as e:
+#     parse_cli = e
 
 
 class Start: