+++ mininterface/settings.py
@@ -6,6 +6,117 @@ from typing import Literal, Optional
 InterfaceName = Literal["gui"] | Literal["tui"] | Literal["textual"] | Literal["text"]
 
 @_dataclass
+class CliSettings:
+    omit_arg_prefixes: bool = False
+    """ Simplify argument names by removing parent field prefixes from flags.
+
+    ```bash
+    $ ./program.py --help
+    # omit_arg_prefixes = False
+    usage: program.py [-h] [-v] --arg.subarg.text STR
+    # omit_arg_prefixes = True
+    usage: program.py [-h] [-v] --text STR
+    ```
+
+    ??? Code
+        ```python
+        @dataclass
+        class SubMessage:
+            text: str
+
+        @dataclass
+        class Message:
+            subarg: SubMessage
+
+        @dataclass
+        class Env:
+            arg: Message
+
+        run(Env, settings=CliSettings(omit_arg_prefixes=True/False))
+        ```
+
+    See: [https://brentyi.github.io/tyro/api/tyro/conf/#tyro.conf.OmitArgPrefixes]()
+    """
+
+    omit_subcommand_prefixes: bool = False
+    """
+    Simplify subcommand names by removing parent field prefixes from subcommands.
+
+    ```bash
+    $ ./program.py --help
+    # omit_subcommand_prefixes = False
+    usage: program.py [-h] [-v] {subcommand:message,subcommand:console}
+    # omit_subcommand_prefixes = True
+    usage: program.py [-h] [-v] {message,console}
+    ```
+
+    ??? Code
+        ```python
+        @dataclass
+        class SubMessage:
+            text: str
+
+        @dataclass
+        class Message:
+            subarg: SubMessage
+
+        @dataclass
+        class Env:
+            subcommand: Message | Console
+
+        run(Env, settings=CliSettings(omit_subcommand_prefixes=True/False))
+        ```
+
+    See: [https://brentyi.github.io/tyro/api/tyro/conf/#tyro.conf.OmitSubcommandPrefixes]()
+    """
+
+    disallow_none: bool = False
+    """ Disallow passing None in via the command-line interface for union types containing None.
+
+    ```bash
+    $ ./program.py --help
+    # disallow_none = False
+    usage: program.py [-h] [-v] [--field {None}|INT]
+    # disallow_none = True
+    usage: program.py [-h] [-v] [--field INT]
+    ```
+
+    ??? Code
+        ```python
+        @dataclass
+        class Env:
+            field: int | None = None
+        run(Env, settings=CliSettings(disallow_none=True/False))
+        ```
+
+    See: [https://brentyi.github.io/tyro/api/tyro/conf/#tyro.conf.DisallowNone]()
+    """
+
+    flag_create_pairs_off: bool = False
+    """ Disable creation of matching flag pairs for boolean types.
+
+    ```bash
+    $ ./program.py --help
+    # flag_create_pairs_off = False
+    usage: program.py [-h] [-v] [--foo | --no-foo]
+    # flag_create_pairs_off = True
+    usage: program.py [-h] [-v] [--foo]
+    ```
+
+    ??? Code
+        ```python
+        @dataclass
+        class Env:
+            foo: bool = False
+        run(Env, settings=CliSettings(flag_create_pairs_off=True/False))
+        ```
+
+    See: [https://brentyi.github.io/tyro/api/tyro/conf/#tyro.conf.FlagCreatePairsOff]()
+    """
+
+
+
+@_dataclass
 class UiSettings:
     toggle_widget: str = "f4"
     """ Shortcut to toggle widgets like secret fields, file picker, and datetime dialog.
@@ -37,9 +148,19 @@ class UiSettings:
 class GuiSettings(UiSettings):
     # If multiple Gui interfaces exist, this had to be TkSettings instead.
 
-
     combobox_since: int = 10
-    """ The threshold to switch from radio buttons to a combobox. """
+    """ The threshold to switch from radio buttons to a combobox.
+
+    Without combobox:
+
+    ![Default](asset/configuration-not-used.avif)
+
+    With combobox:
+
+    ![Combobox](asset/configuration-used.avif)
+
+    (Note, there must be multiple fields for combobox to appear.)
+    """
 
     radio_select_on_focus: bool = False
     """ Select the radio button on focus. Ex. when navigating by arrows. """
@@ -122,5 +243,6 @@ class MininterfaceSettings:
     textual: TextualSettings = _field(default_factory=TextualSettings)
     text: TextSettings = _field(default_factory=TextSettings)
     web: WebSettings = _field(default_factory=WebSettings)
+    cli: CliSettings = _field(default_factory=CliSettings)
     interface: Optional[InterfaceName] = None
     """ Enforce an interface. By default, we choose automatically. """